LayoutTests/inspector/unit-tests/css-keyword-completions-expected.txt

Testing WI.CSSKeywordCompletions.


== Running test suite: WI.CSSKeywordCompletions
-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyName.emptyTextDisallowsEmptyPrefix
PASS: Expected exactly 0 completion results.
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyName.emptyTextAllowsEmptyPrefix
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyName.singleCharacterMatches
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyName.multipleCharacterMatches
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.EmptyTop
PASS: Expected result prefix to be ""
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.EmptyColor
PASS: Expected result prefix to be ""
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.PartialColor
PASS: Expected result prefix to be "papaya"
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.EmptyWordSpacing
PASS: Expected result prefix to be ""
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.MidLineBorder
PASS: Expected result prefix to be ""
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.EnvironmentFunctionCompletion
PASS: Expected result prefix to be ""
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.MidLineEnvironmentFunction
PASS: Expected result prefix to be ""
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.PartialFunctionValue
PASS: Expected result prefix to be "au"
PASS: All expected completions were present.

-- Running test case: WI.CSSKeywordCompletions.forPartialPropertyValue.ComplexMultiValueFunction
PASS: Expected result prefix to be ""
PASS: All expected completions were present.
